[JT-30] refactor: Security 파일 이동
- jwtProvider -> jwtService로 이름 변경
- Member 관련 enum 코드 정리
- 네이버에 맞는 Gender 수정
