+++ module-application/src/main/java/com/devtoon/jtoon/security/filter/JwtAuthenticationFilter.java
@@ -8,9 +8,9 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.filter.OncePerRequestFilter;
 import org.springframework.web.servlet.HandlerExceptionResolver;
 
-import com.devtoon.jtoon.security.jwt.application.JwtProvider;
-import com.devtoon.jtoon.security.jwt.domain.CustomUserDetails;
-import com.devtoon.jtoon.security.jwt.domain.MemberThreadLocal;
+import com.devtoon.jtoon.security.application.JwtService;
+import com.devtoon.jtoon.security.domain.jwt.CustomUserDetails;
+import com.devtoon.jtoon.security.domain.jwt.MemberThreadLocal;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
@@ -24,7 +24,7 @@ import lombok.extern.slf4j.Slf4j;
 public class JwtAuthenticationFilter extends OncePerRequestFilter {
 
 	private final HandlerExceptionResolver handlerExceptionResolver;
-	private final JwtProvider jwtProvider;
+	private final JwtService jwtService;
 
 	@Override
 	protected void doFilterInternal(
@@ -37,7 +37,7 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
 		if (accessToken != null && accessToken.startsWith(BEARER_VALUE)) {
 			try {
 				accessToken = accessToken.split(SPLIT_DATA)[1];
-				if (!jwtProvider.isTokenValid(accessToken)) {
+				if (!jwtService.isTokenValid(accessToken)) {
 					String refreshToken = validateAndGetRefreshToken(request);
 					accessToken = regenerateTokens(refreshToken, response);
 				}
@@ -54,16 +54,16 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
 	private String validateAndGetRefreshToken(HttpServletRequest request) {
 		String refreshToken = request.getHeader(REFRESH_TOKEN_HEADER);
 		refreshToken = refreshToken.split(SPLIT_DATA)[1];
-		jwtProvider.isTokenValid(refreshToken);
-		jwtProvider.verifyRefreshTokenDb(refreshToken);
+		jwtService.isTokenValid(refreshToken);
+		jwtService.verifyRefreshTokenDb(refreshToken);
 
 		return refreshToken;
 	}
 
 	private String regenerateTokens(String refreshToken, HttpServletResponse response) {
-		String newAccessToken = jwtProvider.reGenerateAccessToken(refreshToken);
-		String newRefreshToken = jwtProvider.generateRefreshToken();
-		jwtProvider.updateRefreshTokenDb(newAccessToken, newRefreshToken);
+		String newAccessToken = jwtService.reGenerateAccessToken(refreshToken);
+		String newRefreshToken = jwtService.generateRefreshToken();
+		jwtService.updateRefreshTokenDb(newAccessToken, newRefreshToken);
 		response.setHeader(ACCESS_TOKEN_HEADER, BEARER_VALUE + newAccessToken);
 		response.setHeader(REFRESH_TOKEN_HEADER, BEARER_VALUE + newRefreshToken);
 
@@ -71,7 +71,7 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
 	}
 
 	private void authenticate(String accessToken) {
-		Authentication auth = jwtProvider.getAuthentication(accessToken);
+		Authentication auth = jwtService.getAuthentication(accessToken);
 		CustomUserDetails customUserDetails = (CustomUserDetails)auth.getPrincipal();
 		SecurityContextHolder.getContext().setAuthentication(auth);
 		MemberThreadLocal.setMember(customUserDetails.getMember());
