+++ module-application/src/main/java/com/devtoon/jtoon/security/interceptor/MemberInterceptor.java
@@ -3,7 +3,7 @@ package com.devtoon.jtoon.security.interceptor;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.web.servlet.HandlerInterceptor;
 
-import com.devtoon.jtoon.security.jwt.domain.MemberThreadLocal;
+import com.devtoon.jtoon.security.domain.jwt.MemberThreadLocal;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
