+++ module-domain/src/main/java/com/devtoon/jtoon/member/entity/Gender.java
@@ -10,22 +10,27 @@ import java.util.function.Function;
 import java.util.stream.Collectors;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
+import lombok.Getter;
 
+@Getter
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 public enum Gender {
-	MALE,
-	FEMALE;
 
+	MALE("M"),
+	FEMALE("F"),
+	UNKNOWN("U");
+
+	private String naverGender;
 	private static final Map<String, Gender> GENDER_MAP;
 
 	static {
 		GENDER_MAP = Collections.unmodifiableMap(
 			Arrays.stream(Gender.values())
-				.collect(Collectors.toMap(Enum::name, Function.identity())));
+				.collect(Collectors.toMap(Gender::getNaverGender, Function.identity())));
 	}
 
 	public static Gender from(String gender) {
-		return Optional.ofNullable(GENDER_MAP.get(gender))
+		return Optional.ofNullable(GENDER_MAP.get(gender.toUpperCase()))
 			.orElseThrow(() -> new MemberException(ExceptionCode.MEMBER_GENDER_INVALID_FORMAT));
 	}
 }
