+++ module-domain/src/main/java/com/devtoon/jtoon/member/entity/Member.java
@@ -75,8 +75,8 @@ public class Member extends BaseTimeEntity {
 		this.role = role;
 		this.loginType = loginType;
 	}
-	
-	public void updateLastLogin( ) {
+
+	public void updateLastLogin() {
 		lastLoginDate = LocalDateTime.now();
 	}
 }
