+++ module-domain/src/main/java/com/devtoon/jtoon/member/entity/Role.java
@@ -5,6 +5,7 @@ import lombok.AllArgsConstructor;
 
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 public enum Role {
+	
 	USER,
 	ADMIN
 }
