+++ module-domain/src/main/java/com/devtoon/jtoon/member/entity/Status.java
@@ -5,6 +5,7 @@ import lombok.AllArgsConstructor;
 
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 public enum Status {
+	
 	NORMAL,
 	REST,
 	BLACK
